+++ haystack/preprocessor/utils.py
@@ -5,10 +5,8 @@ import tempfile
 import zipfile
 import gzip
 from pathlib import Path
-from typing import Callable, Dict, List, Optional, Tuple, Union, Generator, Any
+from typing import Callable, Dict, List, Optional, Tuple, Union, Generator
 import json
-from selenium import webdriver
-from urllib.parse import urlparse
 
 from farm.data_handler.utils import http_get
 
@@ -394,161 +392,3 @@ def squad_json_to_jsonl(squad_file: str, output_file: str):
             json.dump(doc, jsonl_file)
             jsonl_file.write("\n")
 
-def write_to_files(urls: Any, driver: Any, output_dir: str, base_url: str = None):
-    """
-    Stores content from urls to files
-
-    :param urls: http addresses whoose content need to be stored in files
-    :type urls: list or str
-    :param drive: seleium webdriver
-    :param output_dir: path of directory where files need to be stored
-    :type: str
-    :param base_url: http address of base url
-    :type: str
-    :return: None
-    """
-    if type(urls) != list:
-        urls = [urls]
-    for link in urls:
-        logger.info(f"writing contents from `{link}`")
-        driver.get(link)
-        el = driver.find_element_by_tag_name('body')
-        text = el.text
-        link_split_values = link.replace('https://','').split('/')
-        file_name = '{}/{}.txt'.format(output_dir, '_'.join(link_split_values))
-        data = {}
-        data['url'] = link
-        if base_url:
-            data['base_url'] = base_url
-        data['text'] = text
-        with open(file_name, 'w') as f:
-            f.write(str(data))
-
-def is_internal_url(base_url: str, sub_link: str):
-    """
-    :param base_url: http address of base url
-    :type base_url: str 
-    :param sub_link: sub url extracted from base url
-    :type sub_link: str
-    :return: bool if sub_link is external url or not
-    """
-    base_url_ = urlparse(base_url)
-    sub_link_ = urlparse(sub_link)
-    return base_url_.scheme == sub_link_.scheme and base_url_.netloc == sub_link_.netloc
-
-def is_inpage_navigation(base_url: str, sub_link: str):
-    """
-    to find whether sub-link is inpage navigation link or not
-
-    :param base_url: http address of base url
-    :type base_url: str 
-    :param sub_link: sub url extracted from base url
-    :type sub_link: str
-    :return: bool if sub-link is inpage navigation or not
-    """
-    base_url_ = urlparse(base_url)
-    sub_link_ = urlparse(sub_link)
-    return base_url_.path == sub_link_.path and base_url_.netloc == sub_link_.netloc
-
-def use_chrome(chrome_driver_path: str):
-    """
-    :param chrome_driver_path: executale file path for chrome driver
-    :return: selenium webdriver
-    """
-    options = webdriver.chrome.options.Options()
-    options.add_argument('--headless')
-    driver = webdriver.Chrome(executable_path=chrome_driver_path, options=options)
-
-    return driver
-
-def use_firefox(firefox_driver_path: str):
-    """
-    :param firefox_driver_path: executale file path for firefox driver
-    :return: selenium webdriver
-    """
-    options = webdriver.firefox.options.Options()
-    options.add_argument('--headless')
-    driver = webdriver.Firefox(executable_path=firefox_driver_path, options=options)
-
-    return driver
-
-def extract_sublinks_from_url(driver: Any, base_url: str, existed_links: List = None):
-    """
-    extracts sub links from url
-
-    :param driver: selenium webdriver
-    :param base_url: http address of base url
-    :type base_url: str
-    :param existed_links: already extracted links
-    :type existed_links: List
-    :return: list of sub-links extracted from base url
-    """
-    driver.get(base_url)
-    a_elements = driver.find_elements_by_tag_name('a')
-    sub_links = set()
-    if not (existed_links and base_url in existed_links):
-        sub_links.add(base_url)
-    for i in a_elements:
-        sub_link = i.get_attribute('href')
-        if not (existed_links and sub_link in existed_links):
-            if is_internal_url(base_url=base_url, sub_link=sub_link) and (not is_inpage_navigation(base_url=base_url, sub_link=sub_link)):
-                sub_links.add(sub_link)
-    return sub_links
-
-
-def fetch_data_from_url(urls: Any, output_dir: str, chrome_driver_path: str = None, firefox_driver_path: str = None, extract_sub_links: bool = True):
-    """
-    takes url/urls as input and write contents to files
-
-    :param urls: list of http addresses or single http address
-    :type urls: str or list (Any)
-    :param chrome_driver_path: chrome webdriver path for selenium
-    :type chrome_driver_path: str
-    :param firefox_driver_path: firefox webdriver path for selenium
-    :type firefox_driver_path: str
-    :param extract_sub_links: whether to extract sub-links from urls or not
-    :type extract_sub_links: bool
-    :return : None
-    """
-    path = Path(output_dir)
-    if not path.exists():
-        path.mkdir(parents=True)
-    
-    is_not_empty = len(list(Path(path).rglob("*"))) > 0
-    if is_not_empty:
-        logger.info(
-            f"Found data stored in `{output_dir}`. Delete this first if you really want to fetch new data."
-        )
-        return False
-    else:
-        if (not chrome_driver_path) and (not firefox_driver_path):
-            logger.info(f"chrome driver executable path or firefox driver executable path need to be specified")
-        else:
-            logger.info(f"Fetching from {urls} to `{output_dir}`")
-
-            if chrome_driver_path:
-                driver = use_chrome(chrome_driver_path=chrome_driver_path)
-            elif firefox_driver_path:
-                driver = use_firefox(firefox_driver_path=firefox_driver_path)
-
-            if type(urls) != list:
-                #extract links from url            
-                if extract_sub_links == True:
-                    sub_links = extract_sublinks_from_url(driver=driver, base_url=urls)
-                    #get sub links from url
-                    write_to_files(sub_links, driver, output_dir=output_dir, base_url=urls)
-                else:
-                    write_to_files(urls, driver, output_dir=output_dir)
-            else:
-                #handle list of urls
-                sub_links = {}
-                if extract_sub_links==True:
-                    for url_ in urls:
-                        existed_links: List = list(sum(list(sub_links.values()), []))
-                        sub_links[url_] = list(extract_sublinks_from_url(driver=driver, base_url=url_, existed_links=existed_links))
-
-                    for url in sub_links:
-                        write_to_files(sub_links[url], driver, output_dir=output_dir, base_url=url)
-                else:
-                    for url_ in urls:
-                        write_to_files(url_, driver, output_dir=output_dir)
